requirements.txt
requests~=2.32.3
pillow~=11.0.0
numpy~=2.2.1